Separate card selectors into clas-specific files
