+++ server/game/CardSelectors/BaseCardSelector.js
@@ -1,0 +1,45 @@
+class BaseCardSelector {
+    constructor(properties) {
+        this.cardCondition = properties.cardCondition;
+        this.cardType = properties.cardType;
+        this.gameAction = properties.gameAction;
+
+        if(!Array.isArray(properties.cardType)) {
+            this.cardType = [properties.cardType];
+        }
+    }
+
+    canTarget(card, context) {
+        return (
+            this.cardType.includes(card.getType()) &&
+            this.cardCondition(card, context) &&
+            card.allowGameAction(this.gameAction)
+        );
+    }
+
+    hasEnoughTargets(context) {
+        return context.game.allCards.any(card => this.canTarget(card, context));
+    }
+
+    defaultActivePromptTitle() {
+        return 'Select characters';
+    }
+
+    automaticFireOnSelect() {
+        return false;
+    }
+
+    wouldExceedLimit(selectedCards, card) { // eslint-disable-line no-unused-vars
+        return false;
+    }
+
+    hasReachedLimit(selectedCards) { // eslint-disable-line no-unused-vars
+        return false;
+    }
+
+    formatSelectParam(cards) {
+        return cards;
+    }
+}
+
+module.exports = BaseCardSelector;
