+++ server/game/CardSelectors/UnlimitedCardSelector.js
@@ -1,0 +1,9 @@
+const BaseCardSelector = require('./BaseCardSelector.js');
+
+class UnlimitedCardSelector extends BaseCardSelector {
+    hasReachedLimit() {
+        return false;
+    }
+}
+
+module.exports = UnlimitedCardSelector;
