+++ server/game/CardSelectors/UpToXCardSelector.js
@@ -1,0 +1,19 @@
+const BaseCardSelector = require('./BaseCardSelector.js');
+
+class UpToXCardSelector extends BaseCardSelector {
+    constructor(numCards, properties) {
+        super(properties);
+
+        this.numCards = numCards;
+    }
+
+    defaultActivePromptTitle() {
+        return this.numCards === 1 ? 'Select a character' : `Select ${this.numCards} characters`;
+    }
+
+    hasReachedLimit(selectedCards) {
+        return selectedCards.length >= this.numCards;
+    }
+}
+
+module.exports = UpToXCardSelector;
